backend/microservices/cognitive-screener/requirements.txt
# Cognitive Screener Service - Python Dependencies

# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# PDF Processing
PyPDF2==3.0.1
python-docx==1.1.0

# OpenAI Integration
openai==1.3.0

# Database
pymongo==4.6.0
redis==5.0.1

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.24.3

# Speech Processing (optional)
SpeechRecognition==3.10.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
